+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: (EVM) Accept transactions with a provided `otherParams.nonce` even if pending transactions exist
+- fixed: Fix Botanix asset display name.
 
 ## 4.55.7 (2025-08-01)